:card_file_box: [#4771] Add data migration to convert price logic rules to normal logic
Converted the current behaviour of price logic rules into assignments of
the price logic variable value, including the fallback behaviour to the
product price if no logic rule matches. This last bit is slightly
nuanced, since changing the price on the product no longer results in
the fallback case being automatically updated along, and that requires a
note in the changelog (!).

At first, I wanted to trigger the logic rules only when the last step
was reached, using trigger_from_step, but I'm unsure that it behaves
correctly if the last step is somehow made not applicable with other
logic rules, so instead these rules are always evaluated. If it's safe,
users can update the auto-migrated rules to optimize performance.

I've also opted to not implement the reverse operation due to the
involved complexity - we should definitely recommend to make database
backups before upgrading to 3.0. We can be a bit bolder than usualy
because it's part of a major version with other breaking changes anyway,
and it should be fine (tm) for the majority of the cases.