:wastebasket: [#4771] 'Disable' support for evaluating price logic rules
We can't effectively remove this codepath yet because it's still used
in the migration tests to ensure the calculated price before and after
migrating has the same outcome, rather than testing the implementation
details/state of the database.

API endpoints and UI are removed, so there should not be a way to be
able to get new price logic rules in the system. We'll also remove
admin access and support for copying them when copying a form.