:white_check_mark: [#4771] Update tests for new price calculation behaviour
There are some slight behaviour changes here that need to be documented,
namely the refusal to fall back to product price if price is derived
from a variable, and the fact we can no longer make some sanity check
assertions in the test method. However, the behaviour is covered by
the main assertion of the test.