:boom: [#3283] Removed objecttype URL to UUID import convertion
The convertion of the objecttype property, of the objects api registration, in the form import has been removed. The objecttype now has to be a UUID and objects_api_group is now required.
The convertion was part of some backwards compatibility to allow the old objects api registration. This has been removed in preparation of OF 3.0