:bug: [#4862] Prevent submissions identifying attributes hashing if they still need cosigning
A submission is only processed/registered when it's cosigned, so we
may definitely not hash the identifying attributes when it's still
pending cosigning.

The logout endpoint can't tell if the cosigner or the original author
is logged in to perform the cosigning, so we must look at the state
of the submission itself.

Backport-of: #4865