:wheelchair: Remove the custom helptext-as-tooltip styling
This aligns the implementation again with default project. The help
text is the only way to specify crucial information/context or
requirements for a given form field, and thus should be visible.

Tooltips serve a different purpose - they can provide additional,
informational, context but may not hide essential information for
accessibility reasons. Unfortunately, the Django admin doesn't have
a mechanism for using both flavours of information.