:recycle: [#4398] Refactor validation error handling to use context
Instead of prop-drilling, make sure to process validation errors at the
locations where the shape/structure is known and rely on the generic
mechanisms to pass/display the validation errors.

This also adds some JSDoc type hinting to the helpers because I keep
getting lost on what the expected shape of data/input is.