:adhesive_bandage: [#4320] Fix crash when rendering PDF again for cosigned submission
The PDF was including the cosign data/component for version 1 of cosign,
which is not compatible at all with v2 (see #4308), causing crashes.

Generally this doesn't matter since the PDF is generated right after
submission is received and before its cosigned, and we don't usually
need to re-generate it except during development.