:recycle: [#4320] Refactor cosign state into dedicated datastructure
The cosign information/state was smeared out across the form and
submission model, and is consulted in a number of places to control
template/feedback behaviour.

The implementation details of how cosign works are now properly
encapsulated in a data structure, which also allows us to prepare for
cosign v3 which will most-likely not be Formio based.