[#3283] Removed warning and updated appointments views
The appointments views used to have a deprecation warning which had to do
with what we return to the frontend. Formio wasn't handling correctly
the raised error of a serializer
(serializer.is_valid(raise_exception=True)) so we had to return an empty
list instead.

Now, we have custome react code and we can make the API calls when we are ready
so we avoid errors, but if any they will now be raised.