:green_heart: Fix flaky msgraph test
The lack of deterministic ordering when processing the submission
attachments can cause tests to sometimes fail in CI.

When no explicit ordering is specified, Postgres figures out an ordering
which tends to correspond to how the tuples/records are stored on disk,
and on particular bad alignment of the stars the later-created record
can appear before earlier ones, being returned in a non-intuitive
order. For this reason, you should always make ordering explicit or
if ordering doesn't matter, write assertions without assuming any
ordering.

In this case, making the order explicit is the simplest fix.