:fire: Delete migration test code
The tests rely too much on the real models, which breaks other PRs that
add additional fields. Refactoring the test code to freeze the price
calculation is not reasonable, because it relies on the logic engine
evaluation, which can't work with historic models and requires access
to the real models and their models.

We'll have to trust that the tests were sufficient when the code
cleanup was initially authored :fingers_crossed: