:recycle: [#4859] Give form state reset control to parent component
Handling the reset of dependent fields in the ObjectTypeSelect
component is too naive - there's no guarantee that the fields must
always be reset, for example with the copy-config-from-registration
button. Clicking that button updates a bunch of form state at once,
and it already manages the dependent fields.

So, providing a callback to handle the reset allows us to handle the
dependent fields still, while having to option to incorporate extra
context.