:recycle: [#4859] Display copy toggle only if there are relevant backends
The toggle is not displayed if there are no backends to copy from, so
the UI can be de-cluttered.

Additionally, the mechanism/slot for the extra controls has been
refactored to be more generic - anything can now be rendered there that
may be relevant, and the state is specific to the UI being rendered so
that has been removed from the generic parent component.

We use the formik status to track form-specific state, through a small
hook that normalizes the shape of the status (since it can be anything).