:bug: [#4884] Ensure that no form variables are created for soft required errors
The soft required errors component should have the same behaviour like
the content component, as it holds no submission values and is purely
client-side behaviour.

Formio itself excludes the content component when looping over a
configuration with eachComponent, so we need to add this same logic
to our own code since Formio is not aware of our custom component
semantics.

In the backend, we extend the behaviour for content components to the
softRequiredErrors component to ensure no variables are created in
the database either.