:boom: [#2177] Changing map data to geoJson geometry
For the map interaction changes, we need to save the map data in a different format. This is to allow us to differentiate between point, line or polygon map information.

At the moment we only need the geometry information of the geoJson, so thats what we will store. The other geoJson information, `type` and `properties`, are at the moment not used. In addition, the backend registrations (objects API, ZGW APIs and StUF-ZDS) also only need the geometry information.