:boom: [#3283] Remove legacy machtigen context
The shape of data that we store was badly documented, not in the least
because we barely had any idea what kind of information we receive.

Since Open Forms 2.7.0 it has been fleshed out what the mandata context
looks like, and properly implemented backed by a proposal for a
standard which is implemented all the way to Open Zaak to try out.

This legacy context is now being removed in favour of the new way
to track and expose the mandata context.