:wastebasket: [#3283] Deprecate legacy OIDC callback endpoints
With Open Forms 3.0, we can make the breaking change of updating the
default.

The new OIDC API endpoints all point to a single callback URL, so it
means less configuration overhead on the identity provider side. The
breaking change means that the new URI's must be added to the
allowlist of the identity provider.

The new endpoint is '/auth/oidc/callback/', and it applies to all
OIDC configuration flavours.