:label: Suppress type errors on decorator_include
It's doing a lot of duck-typing and does not play nice with the types
defined in django-stubs, which would ideally be protocols rather than
named types for these kind of things.

However, we let the decorator_include declare the type it wants in
the path arguments, and apply a cast around the custom class to
silence the remaining type checker warnings.

The most important thing here is that the input types for decorator_include
are checked and our urls.py modules aren't unusable.