:boom: [#3283] Make objects/objecttypes FKs non-nullable in Objects API group
Catalogi and Documents API are optional, but the objecttypes and objects
API services are not. These fields were null because the model was
originally a django-solo singleton model which must be able to be
created without any services existing prior.

A check script is added to prevent being stuck in half-migrated
upgrades.