:recycle: [#3457] Refactor payments module to apply plugin mechanism consistently
All other modules operate on the idea that the generic code populates
the plugin options from the generic information and the specified
options serializer, which allows plugins to be generically typed and
not needing to perform additional input validation or conversion from
the raw JSON data for the options to the Python/Django objects that
are easier to work with.

This does change the public API of the registry/plugins to take an
extra argument, but enables us to apply static type checking (which
already reveals some problems) and simpler plugin code.