:label: [#3457] Extend type checking to most of the payments app
Rather than just type-checking some files, we are close enough to
type check most of the package and instead skip some files/subpackages
with known (but not critical) type checking issues.

Here and there we need to suppress some errors because the Django ORM
magic is too much for Pyright.