:sparkles: [#3457] Support custom templates for Ogone COM and TITLE parameters
The COM parameter is sent/displayed in the payment receipts of the
merchant and can contain references for the organization's
administration.

The TITLE is (supposedly) displayed on the payment page, however this
seems not to be functioning with the simulator.

If no custom templates are specified, we use the legacy behaviour and
include a hardcoded description.