[#4886] Added check for csv uploads to the MimeTypeValidator
There is a problem defining that a file is text/plain or text/csv.

We decided to make some extra checks in order to be closer to that
conclusion because a lot of csv files are not created properly
(according to the proper structure, delimiters etc.) and are treated as
text/plain from magic library.