:bug: [#4795] Validate .msg file type only on backend
The sdk cannot determine which content type belongs to a .msg file. This is because (at least) Linux and MacOS don't know this file type.

To make sure these files can be uploaded, the type property on the FileSerializer is now optional. For .smg files a new rule has been added to the MimeTypeValidator