:sparkles: [#4930] Implement the actual export
Export submission statistics based on the timeline logs.

Note: if the timeline logs are pruned, this affects the exports. It's
up to the users to periodically create these exports and save them
somewhere if they periodically prune log records.

Note 2: filtering on forms only works on new log records, as existing
log records don't have the form ID stored in the structured data.

Submissions that were deleted for which existing log records are
present will display 'unknown' for some columns because the relevant
information has been deleted. Only from 3.0 onwards are we snapshotting
the data required for the exports.