:card_file_box: [#4920] Result of squashmigrations from 0055 to 0069
Manually commented out some of the RunPython to hopefully better
optimize the squashed migration, and added those operations back in
to the squashed migration (mostly the resetting of the default template
translations which is a workaround).

The update_summary_tags in the original migration now imports the
definition from the squashed migration rather than not executing it
at all, as some dev-environments might miss this step if they pull
master not frequently enough.

This migration can be further optimized by hand, see the next commit.