:card_file_box: [#4920] Clean up and squash objects API migrations
The Objects API group config model was moved from the registrations app
to the general purpose contrib app for the Objects API, as it is used
by both prefill and registration plugins. This involved some dark
migrations magic with SeparateDatabaseAndState to move the model
without breaking the existing table or messing with table names.

Now that we're guaranteed existing instances have upgrade to 2.8, we
know that the tables are properly in the right place and the necessary
migrations have been executed, allowing us to turn them into normal
migrations so that fresh installs can take the direct path rather than
the detour, and it allows us to clean up some migration code.

The rest of the migrations in the contrib app are not yet squashed, as
they are primarly operations around RunPython that need to be executed
on 3.0 instances.