:sparkles: [#4650] Support optional ownership checks in objects API prefill
With product prefill, the premise is that the product data is personal
and likely contains sensitive/private data that gives only the person
with a particular BSN access. This requires the person to authenticate
and then their BSN is compared with some property in the retrieved
object.

However, there may be 'public' data used to prefill some variables,
either user defined or even form variables. Multiple people can use
those objects, even without authenticating in the first place.

You can now flag a prefill to not require this kind of ownership check.