:children_crossing: [#4969] Organize configuration fields on base tab in fieldsets
In an attempt to declutter, related fields are now grouped in fieldsets
and especially fields that are optional are collapsed by default.

The fieldsets report a status where relevant to get a quick overview,
e.g. showing that no auth is configured or that the form is indeed
currently active (without being in maintenance mode).