:white_check_mark: Clarify test for prefill from options
The plugin return value determines which values get assigned to which
variables - there is no default mechanism that stores the retrieved
object in the variable that holds the prefill configuration.

This means that for the objects API, the retrieved object is never
persisted to DB, only the mapped variables contained in the
configuration.