:bug: [#4978] Fix accidental HTML escaping in summary PDF/confirmation email
While the component/formatter properly takes care of conditional
escaping by leveraging format_html and friends, the post-processor
was converting the SafeString into a regular string again by doing
string-interpolation for the file names, which leads to the full
result being HTML escaped again.

In HTML mode, the prefix 'attachment' is now dropped, as the markup and
context of the label/field should provide sufficient information and
the 'attachment:' prefix looks odd in combination with the <ul> markup.