:bug: [4699] Optional addressNL postcode and houseNumber
The postcode and houseNumber weren't made `required` or `optional` depending on the `required` state of the addressNL component. These fields have/had only regex validation, and no `allow_blank` or any onther "optional" ruling. This resulted in the postcode and houseNumber fields to be always required.

With this change the `required` and `allow_blank` values for the postcode and houseNumber fields gets set depending on the required state of the over-arching addressNL component. This makes optional addressNL components correctly optional