:bug: [#4908] Revise how attachments are processed
There was a bug with the previous attachment processing, where attachments would get overwritten if multiple were uploaded. Now, an object will be returned with the 'file_name' and 'content' properties in case of a single file component (or 'None' if no file was uploaded). And in case of a multiple files component, it will be a list of these file objects (or an empty list if no file was uploaded). This keeps the data type consistent, and we can give guarantees that an object will always have the file_name and content keys.