:card_file_box: Clear out RunPython operations
The data migrations that needed to be performed on older instances
are now removed, since the upgrade checks guarantee that they
have been executed during upgrades to 3.0.x. For new instances,
these operations were not relevant in the first place.

This allows us to remove migration test code which speeds up the
test suite and leads to less code to maintain.