:sparkles: [#4931] Set up custom admin template for statistics
The list view will render the aggregation results, rather than each
individual log record, so we override the entire result table and
mimic the django markup.

Note that the ordering options are not copied over/supported as this
looks quite complex in django itself. We might at some point maybe
support search fields on form name (or a list filter?) to narrow
down results if it becomes too unwieldly.