:bug: Fix default version handling for Objects API registration
Because the `version` configuration option was absent for freshly
added registration backends, you can't immediately configure them in
the variables table unless you save the form first and then edit it.

Another crash has been observed where doing the above resulted in the
variables table not rendering (crashing), but I wasn't able to
reproduce this. I suspect it's because the variablesMapping config
option was undefined or null which may happen because of toggling
between v1/v2.

A fix worth backporting, but no need for urgent hotfix releases since
you can work around it by saving the form after adding the registration
backend, and then finishing the registration configuration in the
variables table.