:sparkles: [open-formulieren/open-forms-sdk#76] Prefer ESM modules for the SDK if the browser supports it
Using module type scripts makes it possible to opt into the benefits
of bundle splitting where only code is loaded that will actually be
used.

This sets up a wrapper entrypoint/script gated behind the type=module
feature of browsers so it's not executed by browsers that don't
natively support modules yet. An extra advantage is that we can pass
all the necessary configuration via data attributes and process them
in our script, meaning we don't need inline scripts and CSP nonces
anymore for this.

Finally, the SDK ESM module is added to the preload list since we are
guaranteed that we'll need it, and we avoid a waterfall when loading
the entrypoint and then dynamically importing the SDK bundle - those
can be fetched in parallel.