:bug: [#5035] Fix components in repeating groups creating FormVariable records
Components nested in repeating groups are not real variables. The
frontend code showed these in the variables table AND this also
caused them to be PUT to the backend, resulting in crashes when
processing file uploads because there are no saved submission variables
available for them.

Unfortunately, this doesn't appear to fix the problem with too much
keys being included in the submission data during registration, so
it must have a different root cause.