:bug: [#5035] Fix "phantom" properties in legacy objects API
Fix 1:

Components nested in repeating groups are not real variables. The
frontend code showed these in the variables table AND this also
caused them to be PUT to the backend, resulting in crashes when
processing file uploads because there are no saved submission variables
available for them.

Unfortunately, this doesn't appear to fix the problem with too much
keys being included in the submission data during registration, so
it must have a different root cause.

Fix 2:

Avoid the total configuration wrapper mutating the first step configuration

The total configuration wrapper merges the configuration wrapper of
each step into a single object for optimized access to values/
components. It takes the first step and merges the remaining steps into
it. However, this had the unintended side-effect of mutating the config
of the first step, manifesting in the objects API v1 registration with
the json_summary tag which contained extra, unexpected keys in the
submission data of the first step.

Fixed by making a deep copy first to end up with a different instance
that can be safely mutated.

Backport-of: #5043