:green_heart: Fix broken assertion in test
Jan 1st + 1 month equals Feb. 1st, not Jan 31st. Why was this test passing then, you ask?
Well, the PDF also includes the timestamp when it was generated, based on 'now'. The
PR with the fix (and test) was created on Jan 31st, effectively causing the expected
string to be present in the PDF. Until, of course, time passed and it became
February.