:technologist: [#5058] Remove form_variable FK to avoid confusion for developers
This sets up the factory to make sure the underlying FormVariable also
exists for a given SubmissionValueVariable instance, but we *do* need
to handle the case where this integrity is broken during runtime as
records may be gone for whatever reason (and there are tests for this
situation).

Instead of crashing hard, we just log a message and fall back to just
the raw underlying JSON value. The casting to python types is
troublesome even when all information is available and the scope is
too big to solve that here.